refactor: move PID analysis into signal processing
Refactor pid_analysis.rs into the signal processing
framework as the first module. PID now implements
the SignalAnalysis trait and reads all 6 signals in
one ULog pass instead of 3 separate passes.

The old pid_analysis() function remains as a thin
backward-compatible facade.
