+++ crates/converter/src/lib.rs
@@ -3,3 +3,4 @@ pub mod converter;
 pub mod diagnostics;
 pub mod metadata;
 pub mod pid_analysis;
+pub mod signal_processing;
